Unit Correctness and New Standard for Interactions
The code should now handle changes of units correctly. Because of an
apparent probable mistake in the kato paper, the formulation was a mix
between non-dimensional standardized cases and dimensionally correct
cases.

Interactions are now supposed to yield quantities such as energy and
forces on a per unit area basis. The reason is that it is confusing that
changing the resolution of a problem should entail a change in the
interaction potential.

Furthermore, these fixes seem to render scaling of displacements for
systems in order to help minimizers obsolete.